app/requirements/prd.txt
loguru==0.4.1
passlib[bcrypt]
python-dotenv
requests==2.22.0
tqdm==4.41.1
unsync
